requirement.txt
django
djangorestframework
django-filter
#djangorestframework-csv
mysqlclient
libmysqlclient-dev
